Separate signal table into SignalRow component
