+++ src/components/signal-viewer/renderer.tsx
@@ -1,67 +1,34 @@
 import React from 'react';
-import * as vega from 'vega';
 
 import { View } from '../../constants';
-import Table from '../table';
+import SignalRow from '../signal';
 
 import './index.css';
 
 const KEY = '_signals';
-const header = ['Signal', 'Value'];
 
 interface Props {
   view?: View;
-  debugPane: boolean;
 }
 
 export default class SignalViewer extends React.Component<Props> {
   constructor(props) {
     super(props);
   }
-  public componentDidMount() {
-    Object.keys(this.props.view[KEY]).map(signal => {
-      this.props.view.addSignalListener(signal, () => {
-        this.forceUpdate();
-      });
-    });
-  }
-  public componentWillUnmount() {
-    Object.keys(this.props.view[KEY]).map(signal => {
-      this.props.view.removeSignalListener(signal, () => {
-        // Do nothing
-      });
-    });
-  }
-  public getValue(signalKey) {
-    let returnValue = '';
-    const currentValue = this.props.view.signal(signalKey);
-    if (typeof currentValue === 'object') {
-      Object.keys(currentValue).map(value => {
-        returnValue += `${value}: ${currentValue[value]}, `;
-      });
-      return returnValue.slice(0, returnValue.length - 2);
-    }
-    return currentValue;
-  }
-  public getData() {
-    let values = [];
-    Object.keys(this.props.view[KEY]).map((signal, id) => {
-      values = [
-        ...values,
-        {
-          id,
-          [header[0]]: signal,
-          [header[1]]: this.getValue.bind(this)(signal),
-        },
-      ];
-    });
-    return values;
-  }
   public render() {
-    const data = this.getData.bind(this)();
     return (
       <div className="signal-viewer">
-        <Table header={header} data={data} />
+        <table>
+          <tr>
+            <th>Signal</th>
+            <th>Value</th>
+          </tr>
+          <tbody>
+            {Object.keys(this.props.view[KEY]).map((signal, id) => (
+              <SignalRow key={id} signal={signal} view={this.props.view} />
+            ))}
+          </tbody>
+        </table>
       </div>
     );
   }
